chatAI_streamlit_langchain/requirements.txt
python-dotenv
duckduckgo-search
streamlit